# sarc/docs/swagger.yaml
definitions:
  domain.Building:
    properties:
      address:
        type: string
      buildingId:
        type: integer
      buildingName:
        type: string
    type: object
...
  domain.Curriculum:
    properties:
      courseName:
        type: string
      dataFim:
        type: string
      dataInicio:
        type: string
      disciplines:
        items:
          $ref: '#/definitions/domain.Discipline'
        type: array
      id:
        type: integer
    type: object
  domain.Discipline:
    properties:
      bibliography:
        items:
          type: string
        type: array
      credits:
        type: integer
      id:
        type: integer
      name:
        type: string
      program:
        type: string
    type: object
  domain.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /buildings:
    get:
      description: Retrieves all buildings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Building'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get all buildings
      tags:
      - buildings
    post:
      consumes:
      - application/json
      description: Creates a new building in the system
      parameters:
      - description: Building data
        in: body
        name: building
        required: true
        schema:
          $ref: '#/definitions/domain.Building'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Building'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Create a new building
      tags:
      - buildings
  /buildings/{id}:
    delete:
      description: Deletes a building by its ID
      parameters:
      - description: Building ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Delete a building
      tags:
      - buildings
    get:
      description: Retrieves a building by its ID
      parameters:
      - description: Building ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Building'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Building not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get building by ID
      tags:
      - buildings
    put:
      consumes:
      - application/json
      description: Updates the building information for the given building ID
      parameters:
      - description: Building ID
        in: path
        name: id
        required: true
        type: integer
      - description: Building data
        in: body
        name: building
        required: true
        schema:
          $ref: '#/definitions/domain.Building'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Building'
        "400":
          description: Invalid ID or bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Update an existing building
      tags:
      - buildings
...
  /curriculums:
    get:
      description: Retrieves all curriculums
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Curriculum'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get all curriculums
      tags:
      - curriculums
    post:
      consumes:
      - application/json
      description: Creates a new curriculum in the system
      parameters:
      - description: Curriculum data
        in: body
        name: curriculum
        required: true
        schema:
          $ref: '#/definitions/domain.Curriculum'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Curriculum'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Create a new curriculum
      tags:
      - curriculums
  /curriculums/{id}:
    delete:
      description: Deletes a curriculum by its ID
      parameters:
      - description: Curriculum ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Delete a curriculum
      tags:
      - curriculums
    get:
      description: Retrieves a curriculum by its ID
      parameters:
      - description: Curriculum ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Curriculum'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Curriculum not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get curriculum by ID
      tags:
      - curriculums
    put:
      consumes:
      - application/json
      description: Updates the curriculum information for the given curriculum ID
      parameters:
      - description: Curriculum ID
        in: path
        name: id
        required: true
        type: integer
      - description: Curriculum data
        in: body
        name: curriculum
        required: true
        schema:
          $ref: '#/definitions/domain.Curriculum'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Curriculum'
        "400":
          description: Invalid ID or bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Update an existing curriculum
      tags:
      - curriculums
...
swagger: "2.0"